simplified-template/requirements.txt
# Simplified Template Requirements
# Core ADK and A2A dependencies
google-adk[all]>=1.0.0
a2a-sdk>=0.3.0

# Web framework and ASGI server
starlette>=0.46.2
uvicorn[standard]>=0.34.2
fastapi>=0.115.12

# HTTP client
httpx>=0.28.1

# Configuration and data handling
pyyaml>=6.0
pydantic>=2.11.4
python-dotenv>=1.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type checking (optional)
mypy>=1.5.0
types-PyYAML>=6.0.0

# SSL/TLS support (built into Python, but useful for development)
cryptography>=41.0.0